Fix main methods
